A much more extended version-5.0. It includes results on three related research tasks/problems:
1. Quantitative comparative analysis of the differences between reduced register news headlines and their full-sentence canonical versions. This is based on an interactively refined feature-value schema of the differences.

2. Study of transformation patterns and their analysis in different ways. Question is how many rules capture how many transformations between the two registers. Does a rule-based approach capture the differences and to what extent?

3. Cross-register complexity and similarity analysis: Which direction of transformation/translation is more complex? Finding is that H2C is less difficult that C2H.

The most important final (3rd) point here is that the scenario is just like Machine Translation between registers/varieties. The point is to study the differences of complexity/similarity between the two varieties in terms of controlled experiments using conditional language models, albeit of a linguistic analysis kind, not statistical. But they can still be considered information theoretic, because the study is not really based on meaning at all. It is about purely morphosyntactic patterns.

All of this can shed light on the question of language complexity/similarity, as well as on the validation of clearly delineated distinction between morphosyntax and semantics/pragmatics.

The results so far show that two varieties, in terms of morphosyntax alone, can indeed be different in complexity. Moreover, this is verified by the large differences between the two direction of comparisons based on transformations.